# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Cenos API
  description: |
    API do ecossistema Cenos para gestão de agentes de campo (leituristas)
    das concessionárias de energia do Piauí (PI) e Maranhão (MA).

    ## Autenticação

    A API suporta 5 métodos de autenticação:

    1. **JWT Bearer** — Painel administrativo (`Authorization: Bearer <token>`)
    2. **Telegram TMA** — App do agente (`X-Telegram-Init-Data`)
    3. **Token Simples** — Consultas externas/BI (`?token=API_TOKEN`)

    Consulte a documentação detalhada em `docs/` para cada módulo.
  version: "2.0.0"
  contact:
    name: Cenos Suporte
    url: https://cenos.com.br
servers:
  - url: https://api.izi.tec.br
    description: Produção
  - url: http://localhost:3040
    description: Desenvolvimento
...
paths:
...
  /admin/chat/rooms/{roomId}/messages:
    get:
      tags: [Admin, Chat]
      summary: Mensagens da sala
      security:
        - JwtAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Histórico de mensagens
...
  /admin/message_templates:
    get:
      tags: [Admin]
      summary: Lista templates
      security:
        - JwtAuth: []
      responses:
        "200":
          description: Templates
    post:
      tags: [Admin]
      summary: Cria template
      security:
        - JwtAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                text:
                  type: string
                file:
                  type: string
                web_app_button_text:
                  type: string
                web_app_button_url:
                  type: string
      responses:
        "201":
          description: Template criado
  /admin/message_templates/{id}:
    put:
      tags: [Admin]
      summary: Atualiza template
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template atualizado
    delete:
      tags: [Admin]
      summary: Remove template
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template removido

  # ─── ADMIN: SERVICES ─────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticação administrativa via JWT.
        Header: `Authorization: Bearer <token>`
        O token é obtido via `POST /admin/user/login`.
    TelegramAuth:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description: |
        Autenticação do agente via Telegram Mini App.
        Header: `X-Telegram-Init-Data: <initData>`
        O initData é validado via Telegram TMA.
    ApiTokenAuth:
      type: apiKey
      in: query
      name: token
      description: |
        Token simples para consultas externas.
        Query param: `?token=API_TOKEN`
...